Fluent save_results and save_restart

+++ coupling_components/solver_wrappers/python/tube_flow_solver.py
@@ -231,11 +231,11 @@ class SolverWrapperTubeFlow(Component):
         super().finalize()
 
     def output_solution_step(self):
-        if self.save_restart != 0 and self.n % self.save_restart == 0:
+        if self.n > 0 and self.save_restart != 0 and self.n % self.save_restart == 0:
             file_name = join(self.working_directory, f'case_timestep{self.n}.pickle')
             with open(file_name, 'wb') as file:
                 pickle.dump({'a': self.a, 'p': self.p, 'u': self.u}, file)
-            if self.save_restart < 0:
+            if self.save_restart < 0 and self.n + self.save_restart > self.timestep_start:
                 try:
                     os.remove(join(self.working_directory, f'case_timestep{self.n + self.save_restart}.pickle'))
                 except OSError:
